Prediction step completed

+++ components/ml/org.wso2.carbon.ml.core/src/main/java/org/wso2/carbon/ml/core/spark/algorithms/SparkModelUtils.java
@@ -399,9 +399,9 @@ public class SparkModelUtils {
 
     }
 
-    public static double getUnique(String statsAsJson) {
+    public static int getUnique(String statsAsJson) {
         if (statsAsJson == null) {
-            return 0.0;
+            return 0;
         }
         try {
             // new JSONArray(statsAsJson).getJSONObject(0).getJSONArray("values").getJSONArray(0).getString(0)
@@ -409,18 +409,18 @@ public class SparkModelUtils {
             JSONObject jsonObj = array.getJSONObject(0);
             String unique = jsonObj.getString("unique");
             if (unique == null) {
-                return 0.0;
+                return 0;
             } else {
                 try {
-                    return Double.parseDouble(unique);
+                    return Integer.parseInt(unique);
                 } catch (NumberFormatException e) {
-                    return 0.0;
+                    return 0;
                 }
             }
 
         } catch (JSONException e) {
             log.warn("Failed to extract unique values from summary stats: " + statsAsJson, e);
-            return 0.0;
+            return 0;
         }
 
     }
